+++ e2e/hyperlane/config/config.json
@@ -58,10 +58,10 @@
           "http": "http://localneutron-1-val-0-neutronic:9090"
         }
       ],
-      "interchainGasPaymaster": "0x19e487b113825ab35f9bed5f88b37c626221a2c80170e6d46746b84d18e9e0b0",
+      "interchainGasPaymaster": "0xa83b42b893ef7023b40e908aea0e9e35ec70fb661f604aa99bb4e1223bd70a18",
       "isTestnet": false,
-      "mailbox": "0x3d60f7900be67775c64edd9e0c9d65cd252ede00b549536b47d8e09d48f33278",
-      "merkleTreeHook": "0x61369e5c17f13f4a75b22b1349c818c7f9fa624f0409b4a89814d010a6e480b6",
+      "mailbox": "0xcf581d94e692f6dc8af237693e0cf18044360c9ba9db3eb16aafadb7fde9276c",
+      "merkleTreeHook": "0x1afdb26784f35aa104c07e392017b0f46cccbd4044deb9d9a4826792a93653ff",
       "name": "neutron",
       "nativeToken": {
         "decimals": 6,
@@ -87,7 +87,7 @@
       },
       "slip44": 118,
       "technicalStack": "other",
-      "validatorAnnounce": "0x8ba30384cd9631b2c2b13197186ba32d4de14d5a72e86f7465a8535280c1e6ac"
+      "validatorAnnounce": "0x43f314eb08f37ab7c79e477616a65cff8e1408b2d7a0f4dea5c41366fa3ec591"
     }
   },
   "defaultRpcConsensusType": "fallback"
